refactor(drawer): extract custom tooltip to its own component
