+++ src/client/user/components/Drawer/ControlPanel/widgets/LinkOwnershipField.tsx
@@ -1,36 +1,22 @@
 import React, { FunctionComponent, useState } from 'react'
 import {
-  Tooltip,
   createStyles,
   makeStyles,
   useMediaQuery,
   useTheme,
 } from '@material-ui/core'
 
-import PrefixableTextField from '../../../../widgets/PrefixableTextField'
-import TrailingButton from './TrailingButton'
 import ConfigOption, {
   TrailingPosition,
 } from '../../../../widgets/ConfigOption'
-import helpIcon from '../../../../../app/assets/help-icon.svg'
 import { useDrawerState } from '../../index'
 import useShortLink from '../util/shortlink'
+import DrawerTooltip from './DrawerTooltip'
+import PrefixableTextField from '../../../../widgets/PrefixableTextField'
+import TrailingButton from './TrailingButton'
 
-const useStyles = makeStyles((theme) =>
+const useStyles = makeStyles(() =>
   createStyles({
-    drawerTooltip: {
-      // margin: theme.spacing(1.5, 1, 1.5, 1),
-      whiteSpace: 'nowrap',
-      maxWidth: 'unset',
-      [theme.breakpoints.up('md')]: {
-        marginTop: '-12px',
-        padding: '16px',
-      },
-    },
-    ownershipHelpIcon: {
-      width: '14px',
-      verticalAlign: 'middle',
-    },
     regularText: {
       fontWeight: 400,
     },
@@ -41,7 +27,9 @@ type LinkOwnershipFieldProps = {
   closeModal: () => void
 }
 
-const LinkOwnershipField: FunctionComponent<LinkOwnershipFieldProps> = ({ closeModal }) =>  {
+const LinkOwnershipField: FunctionComponent<LinkOwnershipFieldProps> = ({
+  closeModal,
+}) => {
   const classes = useStyles()
   const theme = useTheme()
   const isMobileView = useMediaQuery(theme.breakpoints.down('sm'))
@@ -52,19 +40,10 @@ const LinkOwnershipField: FunctionComponent<LinkOwnershipFieldProps> = ({ closeM
   const linkOwnershipHelp = (
     <>
       Link owner{' '}
-      <Tooltip
+      <DrawerTooltip
         title="Links can only be transferred to an existing Go.gov.sg user"
-        arrow
-        placement="top"
-        classes={{ tooltip: classes.drawerTooltip }}
-      >
-        <img
-          className={classes.ownershipHelpIcon}
-          src={helpIcon}
-          alt="Ownership help"
-          draggable={false}
-        />
-      </Tooltip>
+        imageAltText="Ownership help"
+      />
     </>
   )
